Using the automatic uninstaller
The chocolateyuninstall.ps1 script does not work well with the latest version of the software (2.7.1.1018). When the script is present, Chocolatey removes the WeChat software and then runs the auto uninstaller, resulting in failure; we have to run choco uninstall wechat again to remove the leftovers. After I removed the script, Chocolatey only uses the auto uninstaller. The removal is not longer silent as a GUI may appear.